[Task] Moved index.js into the sources
