+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "restypie",
+  "main": "./src/index",
   "version": "0.0.2",
   "engines": {
     "node": ">=4.2.2"
