+++ src/api.js
@@ -5,7 +5,7 @@
  **********************************************************************************************************************/
 let _ = require('lodash');
 
-let Restypie = require('../');
+let Restypie = require('./');
 
 /***********************************************************************************************************************
  * @namespace Restypie
