+++ src/basic-routes/lib/get-single.js
@@ -5,7 +5,7 @@
  **********************************************************************************************************************/
 let Promise = require('bluebird');
 
-let Restypie = require('../../../');
+let Restypie = require('../../');
 
 /***********************************************************************************************************************
  * @namespace Restypie.BasicRoutes
