+++ src/basic-routes/lib/patch-many.js
@@ -5,7 +5,7 @@
  **********************************************************************************************************************/
 let Promise = require('bluebird');
 
-let Restypie = require('../../../');
+let Restypie = require('../../');
 
 /***********************************************************************************************************************
  * @namespace Restypie.BasicRoutes
