+++ src/fields/lib/date.js
@@ -5,7 +5,7 @@
  **********************************************************************************************************************/
 let _ = require('lodash');
 
-let Restypie = require('../../../');
+let Restypie = require('../../');
 
 const DEFAULT_ISO_REGEX = /\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d\.\d+([+-][0-2]\d:[0-5]\d|Z)/;
 const TIMESTAMP_REGEX = /^\d{1,16}$/;
