+++ src/fields/lib/float.js
@@ -3,7 +3,7 @@
 /***********************************************************************************************************************
  * Dependencies
  **********************************************************************************************************************/
-let Restypie = require('../../../');
+let Restypie = require('../../');
 
 const FLOAT_REGEX = /^-?\d+(\.\d+)?$/;
