+++ src/resources/lib/proxy.js
@@ -3,7 +3,7 @@
 /***********************************************************************************************************************
  * Dependencies
  **********************************************************************************************************************/
-let Restypie = require('../../../');
+let Restypie = require('../../');          
 let _ = require('lodash');
 let UrlValidator = require('valid-url');
