+++ src/resources/lib/sequelize.js
@@ -3,7 +3,7 @@
 /***********************************************************************************************************************
  * Dependencies
  **********************************************************************************************************************/
-let Restypie = require('../../../');
+let Restypie = require('../../');
 let _ = require('lodash');
 
 /***********************************************************************************************************************
