+++ src/rest-errors/index.js
@@ -3,7 +3,7 @@
 /***********************************************************************************************************************
  * Dependencies
  **********************************************************************************************************************/
-let Restypie = require('../../');
+let Restypie = require('../');
 
 /***********************************************************************************************************************
  * @namespace Restypie
